Download the dataset
Noticed that my last try to download the dataset
didn´t go as planned because I ran the cell that
changes directory twice. So the downloaded file
didn´t went where I wanted. Added a cell to
confirm directory that you work in.
Now the dataset are downloaded to the raw folder.