Saved species.pkl
Now I´ve searched through all files if there
was any files that doesn´t have any extension of
an image. All of the files ended with the right
extension.

After that I tried to save all the
species names to a .pkl-file so that I can use
that file when creating the dashboard. But my
first code didn´t find all of the names,
only 26 out of 94. So that problem was documented
in the README before solving the problem.
An iteration was instead creating that collects
all the names in an array. With that solution all
of the 94 species was found and a .pkl with the
correct data was saved.