Barplot showing percentage distribution
Created a barplot that shows the total percentage
of all images through all species. This barplot
will also be used on all sets later to see if
there is a different distribution from the raw
data.

Noticed that 2 of the species had more than the
others and the lowest and highest amount had a
difference of 369 images (581 - 212).