Split data into train/validation/test sets
With this commit I´ve split the images to the sets
so now I can make an augmentation to get a more
accurate learning to the model training.

Also saved the function to display the barplot of
the percentage distribution in a python file that
can be loaded into the notebooks.