Mean image size
With this commit I collected the mean size of the
images in the train directory and saved it as a
.pkl file for further use when augmentate the
images. Also went through and did an avarage
variation image of all species, I don´t know if
that´s necessary. But did it since the malaria
detector walkthrough did it.

Had some problems with loading
tensorflow.keras.preprocessing with my versions so
used Image from PIL instead.